Divide code into more & better-named modules
