+++ src/game/lanes_and_cars/microtraffic/intelligent_acceleration.rs
