+++ src/game/lanes_and_cars/planning/lane_stroke.rs
@@ -4,8 +4,9 @@ use kay::ID;
 use kay::swarm::{Swarm, CreateWith};
 use monet::Thing;
 use core::geometry::{CPath, band_to_thing};
-use super::materialized_reality::BuildableRef;
-use super::super::{Lane, TransferLane, AdvertiseToTransferAndReport};
+use super::super::construction::materialized_reality::BuildableRef;
+use super::super::lane::{Lane, TransferLane};
+use super::super::construction::AdvertiseToTransferAndReport;
 
 #[derive(Compact, Clone)]
 pub struct LaneStroke {
