+++ src/game/lanes_and_cars/rendering/lane_thing_collector.rs
