+++ src/game/lanes_and_cars/rendering/resources/car.rs
