+++ src/game/lanes_and_cars/rendering/resources/traffic_light.rs
