+++ src/game/mod.rs
@@ -5,6 +5,5 @@ pub fn setup() {
 }
 
 pub fn setup_ui() {
-    lanes_and_cars::lane_rendering::setup();
-    lanes_and_cars::planning::current_plan_rendering::setup();
+    lanes_and_cars::setup_ui();
 }
